+++ nomad/api/__init__.py
@@ -21,6 +21,7 @@ from nomad.api.operator import Operator
 from nomad.api.regions import Regions
 from nomad.api.scaling import Scaling
 from nomad.api.sentinel import Sentinel
+from nomad.api.search import Search
 from nomad.api.status import Status
 from nomad.api.system import System
 from nomad.api.validate import Validate