+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Add scaling policy endpoint (#136)
 * Drop Python 3.5 support
 * Up `requests` lib version 
+* Add support for /search endpoint (#134)